cacher/afficher certains boutons en fonction des permissions

+++ app/resources/views/livewire/user-list.blade.php
@@ -37,10 +37,11 @@
                                 </td>
                                 <td>
                                     <a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm">Modifier</a>
-                                
-                                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                                    {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
-                                    {!! Form::close() !!}
+                                    @can('users.destroy')
+                                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
+                                        {!! Form::close() !!}
+                                    @endcan
                                 </td>
                                 <td>
                                 @can('changepasswd.allusers')
@@ -55,11 +56,15 @@
 
                             @elseif ($mode == "transformation")
                                 <td>
-                                    <a href="{{ route('users.choisirfonction', $user->id) }}" class="btn btn-info btn-sm">Attribuer des fonctions</a>
+                                    @can('users.attribuerfonction')
+                                        <a href="{{ route('users.choisirfonction', $user->id) }}" class="btn btn-info btn-sm">Attribuer des fonctions</a>
+                                    @endcan
                                     <a href="{{ route('transformation.livret', $user->id) }}" class="btn btn-warning btn-sm">Livret de transformation</a>
                                     <a href="{{ route('transformation.progression', $user->id) }}" class="btn btn-primary btn-sm">Progression</a>
                                     <a href="{{ route('transformation.fichebilan', $user->id) }}" class="btn btn-secondary btn-sm">Fiche bilan</a>
-                                    <a href="{{ route('stages.consulter', [ 'marin' => $user->id]) }}" class="btn btn-danger btn-sm">Stages</a>
+                                    @can('stages.consulter')
+                                        <a href="{{ route('stages.consulter', [ 'marin' => $user->id]) }}" class="btn btn-danger btn-sm">Stages</a>
+                                    @endcan
                                 </td>
                                 <td></td>
                             @endif
